+++ CHANGELOG.md
@@ -6,6 +6,8 @@
     - All configured databases can be reloaded in the background
       using `Geolix.reload_databases/0`
     - Database configuration has been extended to support adapters
+    - `Geolix.Adapter.Fake` is provided for a custom managed database with
+      fixed/pre-defined responses
     - "GeoIP-Enterprise" databases return structs instead of plain maps
     - Lookups are done with a configurable timeout